src/test/java/uk/co/pete_b/advent/aoc2018/Day12Test.java
package uk.co.pete_b.advent.aoc2018;

import org.apache.commons.io.IOUtils;
import org.junit.Test;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.io.IOException;
import java.nio.charset.Charset;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.assertEquals;

public class Day12Test {
    private static final Logger LOGGER = LoggerFactory.getLogger(Day12Test.class);

    private static final List<String> TEST_INPUT = Arrays.asList("initial state: #..#.#..##......###...###",
            "",
            "...## => #",
            "..#.. => #",
            ".#... => #",
            ".#.#. => #",
            ".#.## => #",
            ".##.. => #",
            ".#### => #",
            "#.#.# => #",
            "#.### => #",
            "##.#. => #",
            "##.## => #",
            "###.. => #",
            "###.# => #",
            "####. => #");

    @Test
    public void testExamplesPart1() {
        assertEquals(325, Day12.calculateTotal(TEST_INPUT, 20));
    }

    @Test
    public void getAnswers() throws IOException {
        final List<String> input = IOUtils.readLines(getClass().getResourceAsStream("/puzzle-data/2018/day12"), Charset.defaultCharset());
        LOGGER.info("Part 1: " + Day12.calculateTotal(input, 20));
        LOGGER.info("Part 2: " + Day12.calculateTotal(input, 50000000000L));
    }
}
